+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The project follows [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+- Add `npm run release` to prepare the next stable version and changelog.
 - Publish new package versions automatically after successful CI for a push to `master`.
 
 ## [0.1.0] - 2026-07-29